Split PayloadAttribute calculation out of ForkChoiceNotifier
Starting to get a clearer separation of duties instead of ForkChoiceNotifier doing so much itself.
